twist is looking pretty bad. I did a number of things badly. It would be better to start over.
before printing the entire fractal, It might be smarter to simply generate a fractal that is rotated the correct amount for each iteration and then copy and past the same one thousands of times.
This would be FAR, FAR more efficient than simply calculating the rotated coordinates every single time I want to print one. This would be vastly faster. and the benefits would grow exponentially as the iterations increase.