+ ideas for fractal map data structure (images).
Also added random functions (rand.h, rand.c, mt19937 stuff).
Got the first revision of the mapBlockData structure done.
At the moment, the program will not do any SDL stuff. It just fills a single mapBlock with random data between -10 and 20 and then prints it to file.
This is good for evaluating that everything is working as it should be.